Simplify variant system and clean up type hierarchy (#66)
* move paper static files

* Simplify variant system: derive categories from module_path, use node indices in JSON

- Replace hardcoded `categorize_type` and `compute_doc_path` with derivation
  from `ProblemSchemaEntry.module_path` via inventory lookup
- Simplify reduction_graph.json edges to use node indices instead of full objects
- Use `G::NAME` instead of `short_type_name::<G>()` in variant() for correct
  graph type names (e.g. "GridGraph" not "GridGraph<i32>")
- Remove manual `register_types` function; use `Problem::NAME` directly
- Add MIS GridGraph/UnitDiskGraph reductions via unitdiskmapping module
- Update docs (introduction.md, reductions.typ) to dereference node indices


* chore: apply rustfmt formatting


* Remove hardcoded fallback from #[reduction] macro, make ReduceTo impls concrete

The proc macro had a fallback path for type-generic impls that silently
hardcoded "SimpleGraph"/"Unweighted" defaults and used heuristics,
producing wrong variants (e.g., KColoring lost "k" field, MaximumSetPacking
got spurious "graph" field).

Changes:
- Macro: replace type-generic fallback with compile error, remove
  source_graph/target_graph/source_weighted/target_weighted attributes,
  delete dead helper functions
- Make 10 ReduceTo impls concrete across 7 reduction files (i32 or f64)
- MaximumSetPacking→QUBO: only f64 impl (i32 promotes via natural edge)
- Add ConcreteVariantEntry registrations for Unweighted variants so
  natural weight-promotion edges work correctly


* Add design doc: type system cleanup (WeightElement, One, SatisfactionProblem)


* Add implementation plan: type system cleanup (WeightElement, One, SatisfactionProblem)


* feat: add WeightElement trait and One type, remove Unweighted/Weights/NumericWeight


* refactor: update graph problem impls to use WeightElement

Replace verbose trait bounds (Clone + Default + PartialOrd + Num + Zero
+ AddAssign + 'static) with WeightElement in Problem and
OptimizationProblem impls for all 8 graph problems. Use W::Sum as the
metric/value type and W::to_sum() for weight accumulation. Also update
5 reduction rule files that reference these graph problems.


* refactor: update set and optimization problem impls to use WeightElement


* refactor: update reduction rule bounds to use WeightElement


* refactor: rename Unweighted to One in variant metadata


* feat: add SatisfactionProblem marker trait


* chore: cleanup remaining Unweighted references in comments


* update example

* feat: add Triangular graph type and MIS reduction via triangular mapping

Introduces a Triangular lattice graph type (subtype of UnitDiskGraph) and
a reduction from MIS<SimpleGraph> to MIS<Triangular> using the weighted
triangular unit disk mapping, with O(n²) overhead.


* fix: resolve PR review comments and improve test coverage

- Fix variantDefaults.weight from "Unweighted" to "One" in docs
- Update design.md weight hierarchy docs to use "One" instead of "Unweighted"
- Format variants.rs inventory::submit! as multi-line blocks
- Apply rustfmt to sat_ksat.rs macro and other files
- Add UnitDiskGraph → GridGraph reduction test for coverage
- Add Triangular Graph trait method tests for coverage


* refactor: remove manual variant registration, infer natural edges only

Remove ConcreteVariantEntry and variants.rs — reduction graph nodes now
come exclusively from explicit #[reduction] registrations. Natural edges
between same-name variants are still inferred from the subtype partial
order on graph/weight types.


* fix: rustdoc warning for unclosed HTML tags and update CLAUDE.md

- Wrap generic types in backticks in maximumsetpacking_qubo.rs doc comment
- Update CLAUDE.md: Unweighted→One, add Triangular/WeightElement, fix variant ID docs


* docs: add Triangular to design.md graph types and hierarchy

- Add Triangular to graph types table and hierarchy diagram
- Update variant discovery description to reflect automatic node inference


* docs: add natural edge example for Triangular MIS reduction


* test: add natural edge test for MIS/Triangular → MIS/SimpleGraph


* feat: add generic natural-edge reduction system

Introduce GraphCast trait, ReductionAutoCast struct, and
impl_natural_reduction! macro for declarative graph subtype
relaxation reductions. First usage: MIS<Triangular> → MIS<SimpleGraph>.


* test: use Petersen graph and ILP solver for natural edge closed-loop test


---------

+++ examples/reduction_maximumsetpacking_to_qubo.rs
@@ -44,7 +44,7 @@ pub fn run() {
         vec![1, 3, 5], // S4 (overlaps S0, S1, S2)
         vec![0, 4, 7], // S5 (overlaps S0, S1, S3)
     ];
-    let sp = MaximumSetPacking::<i32>::new(sets.clone());
+    let sp = MaximumSetPacking::<f64>::new(sets.clone());
 
     // Reduce to QUBO
     let reduction = ReduceTo::<QUBO>::reduce_to(&sp);
