+++ src/unit_tests/rules/maximumsetpacking_qubo.rs
@@ -7,7 +7,7 @@ fn test_setpacking_to_qubo_closed_loop() {
     // 3 sets: {0,2}, {1,2}, {0,3}
     // Overlaps: (0,1) share element 2, (0,2) share element 0
     // Max packing: sets 1 and 2 → {1,2} and {0,3} (no overlap)
-    let sp = MaximumSetPacking::<i32>::new(vec![vec![0, 2], vec![1, 2], vec![0, 3]]);
+    let sp = MaximumSetPacking::<f64>::new(vec![vec![0, 2], vec![1, 2], vec![0, 3]]);
     let reduction = ReduceTo::<QUBO<f64>>::reduce_to(&sp);
     let qubo = reduction.target_problem();
 
@@ -24,7 +24,7 @@ fn test_setpacking_to_qubo_closed_loop() {
 #[test]
 fn test_setpacking_to_qubo_disjoint() {
     // Disjoint sets: all can be packed
-    let sp = MaximumSetPacking::<i32>::new(vec![vec![0, 1], vec![2, 3], vec![4]]);
+    let sp = MaximumSetPacking::<f64>::new(vec![vec![0, 1], vec![2, 3], vec![4]]);
     let reduction = ReduceTo::<QUBO<f64>>::reduce_to(&sp);
     let qubo = reduction.target_problem();
 
@@ -42,7 +42,7 @@ fn test_setpacking_to_qubo_disjoint() {
 #[test]
 fn test_setpacking_to_qubo_all_overlap() {
     // All sets overlap: only 1 can be selected
-    let sp = MaximumSetPacking::<i32>::new(vec![vec![0, 1], vec![0, 2], vec![0, 3]]);
+    let sp = MaximumSetPacking::<f64>::new(vec![vec![0, 1], vec![0, 2], vec![0, 3]]);
     let reduction = ReduceTo::<QUBO<f64>>::reduce_to(&sp);
     let qubo = reduction.target_problem();
 
@@ -58,7 +58,7 @@ fn test_setpacking_to_qubo_all_overlap() {
 
 #[test]
 fn test_setpacking_to_qubo_structure() {
-    let sp = MaximumSetPacking::<i32>::new(vec![vec![0, 2], vec![1, 2], vec![0, 3]]);
+    let sp = MaximumSetPacking::<f64>::new(vec![vec![0, 2], vec![1, 2], vec![0, 3]]);
     let reduction = ReduceTo::<QUBO<f64>>::reduce_to(&sp);
     let qubo = reduction.target_problem();
